ExoticShoes/requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
Markdown==3.4.3
mysqlclient==2.2.0
Pillow==10.0.0
PyJWT==1.7.1
python-dotenv==1.0.0
pytz==2023.3
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
tzdata==2023.3
